+++ lib/pages/bases/bases.dart
@@ -1,1 +1,3 @@
 export 'quill_description_page.dart';
+export 'setting_page.dart';
+export 'help_page.dart';
