+++ lib/pages/bases/help_page.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+import 'package:stacked/stacked.dart';
+
+import 'package:insoblok/providers/providers.dart';
+
+class HelpPage extends StatelessWidget {
+  const HelpPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return ViewModelBuilder<HelpProvider>.reactive(
+      viewModelBuilder: () => HelpProvider(),
+      onViewModelReady: (viewModel) => viewModel.init(context),
+      builder: (context, viewModel, _) {
+        return Scaffold();
+      },
+    );
+  }
+}
