+++ lib/pages/bases/setting_page.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+import 'package:stacked/stacked.dart';
+
+import 'package:insoblok/providers/providers.dart';
+
+class SettingPage extends StatelessWidget {
+  const SettingPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return ViewModelBuilder<SettingProvider>.reactive(
+      viewModelBuilder: () => SettingProvider(),
+      onViewModelReady: (viewModel) => viewModel.init(context),
+      builder: (context, viewModel, _) {
+        return Scaffold();
+      },
+    );
+  }
+}
