+++ lib/pages/profiles/bookmark_page.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+import 'package:stacked/stacked.dart';
+
+import 'package:insoblok/providers/providers.dart';
+
+class BookmarkPage extends StatelessWidget {
+  const BookmarkPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return ViewModelBuilder<BookmarkProvider>.reactive(
+      viewModelBuilder: () => BookmarkProvider(),
+      onViewModelReady: (viewModel) => viewModel.init(context),
+      builder: (context, viewModel, _) {
+        return Scaffold();
+      },
+    );
+  }
+}
