+++ lib/pages/profiles/profiles.dart
@@ -1,1 +1,6 @@
+export 'avatar_page.dart';
 export 'account_page.dart';
+export 'list_page.dart';
+export 'topic_page.dart';
+export 'bookmark_page.dart';
+export 'moment_page.dart';
