+++ lib/pages/profiles/topic_page.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+import 'package:stacked/stacked.dart';
+
+import 'package:insoblok/providers/providers.dart';
+
+class TopicPage extends StatelessWidget {
+  const TopicPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return ViewModelBuilder<TopicProvider>.reactive(
+      viewModelBuilder: () => TopicProvider(),
+      onViewModelReady: (viewModel) => viewModel.init(context),
+      builder: (context, viewModel, _) {
+        return Scaffold();
+      },
+    );
+  }
+}
