+++ lib/providers/bases/bases.dart
@@ -1,1 +1,3 @@
 export 'quill_description_provider.dart';
+export 'setting_provider.dart';
+export 'help_provider.dart';
