+++ lib/providers/bases/help_provider.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+import 'package:insoblok/utils/utils.dart';
+
+class HelpProvider extends InSoBlokViewModel {
+  late BuildContext _context;
+  BuildContext get context => _context;
+  set context(BuildContext context) {
+    _context = context;
+    notifyListeners();
+  }
+
+  Future<void> init(BuildContext context) async {
+    this.context = context;
+  }
+}
