+++ lib/providers/profiles/profiles.dart
@@ -1,0 +1,6 @@
+export 'avatar_provider.dart';
+export 'account_provider.dart';
+export 'list_provider.dart';
+export 'topic_provider.dart';
+export 'bookmark_provider.dart';
+export 'moment_provider.dart';
