+++ lib/providers/providers.dart
@@ -6,3 +6,4 @@ export 'models/models.dart';
 export 'stories/stories.dart';
 export 'widgets/widgets.dart';
 export 'bases/bases.dart';
+export 'profiles/profiles.dart';
