+++ lib/routers/router.dart
@@ -25,8 +25,28 @@ class Routers {
     _pushReplacement(context, kRouterMain, null);
   }
 
-  static goToAccountPage(BuildContext context) {
-    return _pushToRoute(context, kRouterAccount, null);
+  static goToAccountPage(BuildContext context, {UserModel? user}) {
+    return _pushToRoute(context, kRouterAccount, user);
+  }
+
+  static goToAccountAvatarPage(BuildContext context) {
+    return _pushToRoute(context, kRouterAccountAvatar, null);
+  }
+
+  static goToAccountListPage(BuildContext context) {
+    return _pushToRoute(context, kRouterAccountList, null);
+  }
+
+  static goToAccountTopicPage(BuildContext context) {
+    return _pushToRoute(context, kRouterAccountTopic, null);
+  }
+
+  static goToAccountBookmarkPage(BuildContext context) {
+    return _pushToRoute(context, kRouterAccountBookmark, null);
+  }
+
+  static goToAccountMomentPage(BuildContext context) {
+    return _pushToRoute(context, kRouterAccountMoment, null);
   }
 
   static goToCreateRoomPage(BuildContext context) {
@@ -51,5 +71,13 @@ class Routers {
 
   static goToQuillDescriptionPage(BuildContext context, {String? origin}) {
     return _pushToRoute(context, kRouterQuillDescription, origin);
+  }
+
+  static goToSettingPrivacyPage(BuildContext context) {
+    return _pushToRoute(context, kRouterSettingPrivacy, null);
+  }
+
+  static goToHelpCenterPage(BuildContext context) {
+    return _pushToRoute(context, kRouterHelpCenter, null);
   }
 }
